requirements.txt
Flask
flask_sqlalchemy==3.1.1
lmqg==0.1.1
numpy==1.26.4
pandas==1.4.2
Requests==2.31.0
spacy==3.7.4
Werkzeug==2.2.2
tiktoken
langchain
openai
weaviate-client
scikit-learn
fuzzywuzzy
python-Levenshtein
torch
torchvision
faiss-cpu
Pillow
